requirements.txt
astroid==2.3.3
autopep8==1.4.4
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==2.2
django-allauth==0.41.0
django-countries==5.5
django-filter==2.2.0
djangorestframework==3.11.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
pep8==1.7.1
Pillow==7.0.0
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pylint==2.4.4
python3-openid==3.1.0
pytz==2018.5
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.2.4
stripe==2.42.0
typed-ast==1.4.0
urllib3==1.25.7
wrapt==1.11.2
